Backend-Flask/requirements.txt
flask
torch
torchvision
pillow
numpy
flask-cors
